requirements.txt
pyserial
simple_rpc
arduino-simple-rpc
uvicorn
fastapi

